requirement.txt
Flask
requests
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
python-dotenv
bcrypt
psycopg2-binary
google-cloud-speech
google-cloud-translate
elevenlabs
pytube
yt-dlp
google-cloud-storage
google-cloud-texttospeech